+++ ZMPT101B-lib/CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [1.0.2404]
+### Changed
+- Changed to use __Expected\<ZMPT101B_ACVoltage\>__ as result of RMS readers.
+  
 ## [1.0.2203]
 ### Changed
 - Fixed issue with disposing resources.